// uvm_ref_flow_1.1/.dvt/ai/@api/v9.d.ts
export type Version = {
    version: 9
};

/**
 * Represents any type of message that can appear in a conversation.
 */
export type MessagePart = SystemMessagePart | UserMessagePart | AssistantMessagePart | ToolMessagePart;

/**
 * System prompt message.
 * IMPORTANT: The system prompt can only be set via {@link Profile.systemPrompt}.
 */
export type SystemMessagePart = {
    role: 'system';
    content: string;
};

/**
 * Represents any message generated by the assistant.
 * Can either be plain text or a tool call.
 */
export type AssistantMessagePart = AssistantTextMessagePart | AssistantToolCallMessagePart;

/**
 * Message from the user.
 */
export type UserMessagePart = {
    role: 'user';
    content: {
        type: 'text';
        text: string;
    };
};
// ...
export type ToolMessagePart = {
    role: 'tool';
    toolResult: ToolResult;
    toolCall: ToolCall;
};
// ...
export type Options = {
    [k: string]: any;
};
// ...
export type ModelInfo = LanguageModelInfo | EmbeddingModelInfo;

export type DefaultModelConfig = Omit<DynamicModelConfig, 'provider'>;

export type ModelConfig = {
    provider: string;
    providerOptions?: Options;
    model: string;
    modelOptions?: Options;
    modelInfo?: ModelInfo;
};
// ...
export type SessionBlueprintBase = {
    /**
     * Represents the canonical name or id of this blueprint.
     *
     * IMPORTANT: Using a name that is already registered by existing user-defined or built-in blueprints, will overwrite those blueprints with this implementation.
     */
    name: string;

    /**
     * A human-readable string which is rendered in the UI.
     */
    description?: string;

    /**
     * Represents the functionality which uses the blueprint. Used to differentiate the blueprint type.
     */
    target: string;

    /**
     * The canonical name or id of the used profile. If the specified profile is not found, a default one will be used.
     */
    profile: string;
} & Version;

export type ChatSessionBlueprint = SessionBlueprintBase & {
    target: 'chat';

    /**
     * A list of alternating user / assistant message parts used to initiate the conversation
     * when starting a new chat session using this blueprint.
     *
     * IMPORTANT: The last message part should be a user message part.
     */

    messages: MessagePart[] | ((modelConfig: ModelConfig, abortSignal?: AbortSignal) => Promise<MessagePart[]>);
};